requirements.txt
fetchai-sdk
python-dotenv
flask
gunicorn
langchain
langchain-openai
tavily-python